refactor: LinkedAvatarコンポーネントをリファクタ
